[Refactor] : TransportDTO jsondecode name 변경
Git : #10
